Removing 10.2 codecs to avoid backward compatibility issues during Lucene upgrades (NVIDIA/cuvs-lucene#29)
As per [last night's comment on the Solr Jira issue](https://issues.apache.org/jira/browse/SOLR-17892?focusedCommentId=18024203&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-18024203) we realized that exposing the Lucene Codecs in the SPI registry prevents any calling application from being able to upgrade the codec because, under the hood, the next version of Lucene moves the older codec (Lucene101) into a different package name. 
As a result, we should only provide the relevant vector format, reader, and writer, and the calling application needs to wrap them into its own FilterCodec. 

For the sake of the current tests, we have decided to move the codec classes into the tests so that they are not bundled in the artifacts. 

This issue is a blocker for the inclusion of cuvs Lucene in Solr or any other Lucene-based search engine.

Authors:
  - Vivek Narang (https://github.com/narangvivek10)
  - Ishan Chattopadhyaya (https://github.com/chatman)

Approvers:
  - Ishan Chattopadhyaya (https://github.com/chatman)
  - MithunR (https://github.com/mythrocks)

URL: https://github.com/rapidsai/cuvs-lucene/pull/29